requirements.txt
appdirs==1.4.4
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cssselect==1.3.0
curl_cffi==0.13.0
dash==3.2.0
fake-useragent==2.2.0
feedparser==6.0.12
Flask==3.1.2
frozendict==2.4.6
gunicorn==23.0.0
html5lib==1.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
multitasking==0.0.12
narwhals==2.10.2
nest-asyncio==1.6.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parse==1.20.2
patsy==1.0.2
peewee==3.18.3
platformdirs==4.5.0
plotly==6.4.0
protobuf==6.33.0
pycparser==2.23
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-futures==1.0.2
requests-html==0.10.0
retrying==1.4.2
scipy==1.16.3
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
statsmodels==0.14.5
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
w3lib==2.3.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
yahoo-fin==0.8.9.1
yahooquery==2.4.1
yfinance==0.2.66
zipp==3.23.0
